A00826147-homework-02
./chown <username>
Requires 1 argument (valid linux username)
This executable requires sudo/root to retrieve the uid form the user.
The file group will not be modified.
For this homework, this will modify "file.txt" and change the ownership to the supplied username.

To build the executable cd into labs/02/chown_src/ and make

Signed-off-by: PatoGuereque <contacto@patoguereque.com>